Unrolled schedule option
Had to wrap unrolled and vectorized for loops in a trivial for loop so that the var name serves as a marker for scheduling. The for loop goes away when you constant fold later. Seems like a hack - maybe we should be able to mark statements in some other way?