Refactor of reductions and lowering
Introduce the Provide stmt to defer stores until later in lowering.

Reductions are now split into the init and update entries in the schedule so that their bounds can be specified separately. The domain of the update is part of the body of the reduction function. Expressing these bounds from C++ is done using the new RVar type. No inference of RVar bounds is done.