First pass at separating storage from computation
Pipeline node now does not allocate. That's handled by the new
Allocate node. Chunk now takes a pair of dimensions - the level
to store at and the level to compute at:

buffer_t's now separately track size and strides. They also can
express min, which defines what function index is meant by the
first element of the buffer.

Outstanding bugs:
ptx is broken
dense loads are broken because innermost stride is unknown

To do before this is useful performance-wise:
Only compute the new values that are needed when you're inside
a serial for loop. I.e. do a rectangle exclusion of the stuff
computed on all previous iterations.