IR node make methods now return Stmt and Expr instead of pointers
All uses of them cast them to Stmt or Expr immediately anyway, and
this lets us do more interesting eager simplification.

E.g:
Add::make(foo, 0) can return foo, which is possibly not an Add node.

TBD to what extent this is a good idea, as it could just slow things
down with the extra checks. However, this change (or the previous one)
does not seem to affect compile times at all.


Former-commit-id: b2528589f0e9aabaa5a764c15cff97e1dba788fb