ptx backend memory leak fixes
Now cleans up the cuda module, events, and context when the
JITCompiledModule dies. There was also a bug with CodeGen_PTX_Dev
always creating a module but only setting owns_module when it hits
a ptx kernel.

Valgrind now reports 0 bytes definitely or indirectly lost, but there
are still a number of "possibly lost" blocks in libcuda.


Former-commit-id: 896088a05460dab2e806295c669064ad15eaee46