Fix find_inverse.c so that it generates a correct table.
For signed i32 division by 209, it was inaccurately accepting a multiply
by 1315205297 and a shift by 6 for that, even though the correct result
is 2630410593 and 7.   Now, it does 10x more random samples of values
and generates the right result.

(The checked in table is correct, though; presumably it was generated
with a different version of find_inverse.  Sending this patch in to
avoid trouble in the future.)