When splitting, instead of rounding up, clamp the base index.
I.e. if you need to evaluate 6 elements vectorized by 4, instead of
computing [0 1 2 3] [4 5 6 7], compute [0 1 2 3] [2 3 4 5].

Instead of buffers needing to be a multiple of the split factor, they
now need to be at least the split factor. If they're less, it expands
the bounds in the downwards direction. I.e. for three elements, the above
would compute elements [-1 0 1 2].

This behavior doesn't trigger (and thus incurs no overhead) if you statically
promise that the bounds are a multiple of the split factor.


Former-commit-id: 10a424e9a72f1f14ce12fcb73230b9f466800a12