Refactored IR, and made tracing good.
Remove the PrintStmt IR Node. Added String constants, IfThenElse
blocks, and Evaluate, which evaluates its Expr arg and discards it.

The IfThenElse allowed for removing the maybe_ intrinsics.

Tracing can now dump in a binary format to a file, call a custom
function at runtime, or print ascii to the terminal. It can be enabled
or disabled for each Func with fine granularity.


Former-commit-id: 6ca1255f619ab7c2c8ef2714f5b69766e2d1399e