Continuing to mess with bounds inference to fix autotuning bugs
We're not there yet, but this checkpoint is only a little bit broken
and only a little bit slow, and I believe it's better than what's currently
checked in.

HL_TARGET=ptx interpolate is slow because storage inference operates
inside-out and considers both calls and provides.

Some schedules may cause uncaught out-of-bounds reads on inputs, because
the current bounds inference logic does something awkwardly in between
checking what will be loaded and checking what is necessary for a correct
output.