Add a pass that dynamically skips stages (produce/update nodes).
We can add IfThenElses to skip stages where all uses are guarded by a
select on a constant-for-the-lifetime-of-this-realization condition. Values
from these realizations will still be loaded (e.g. due to vector selects),
so the allocations remain, but the values stored there no longer need to be
correct (or even initialized).

This buys you computing something only at some pixels (e.g. sift descriptor).
See the lazy_convolution test for that use case.

It also can turn off entire pipeline stages. See the skip_stages test for
that usage.

It can also turn things on and off at tile granularity, but this relies
on the simplifier being able to detect that the indices within the tiles
really don't affect the condition, and that the condition is therefore
constant across the tile. This is hard and may fail in non-trivial cases.
See process_some_tiles for this use case.