Overhaul of binary tracing to support post-mortem trace analysis.
Fixed bug causing trace files to be ridiculously huge. They're still large.

Every trace event has a parent id. The trace function returns an integer to
use as a parent id for child events. This helps disambiguate loads/stores
to different realizations of the same func occuring in different threads.

There's the start of a trace analysis tool checked in.


Former-commit-id: 13e2bfca6013295739ec9806177155361f4ace71