Remove unnecessary inline reduction name prefixing.
We ended up handling bounds for multiple instances of one func differently.
Instead of making copies during bounds inference, each stage tracks a map
from consumers to boxes, and merges the appropriate ones depending on which
consumers are in scope when it's time to lay down some bounds inference code.