Remove unnecessary inline reduction name prefixing.
We ended up handling bounds for multiple instances of one func differently.
Instead of making copies during bounds inference, each stage tracks a map
from consumers to boxes, and merges the appropriate ones depending on which
consumers are in scope when it's time to lay down some bounds inference code.


Former-commit-id: a51bac29d3dc0d22e286e4d81e9b968e68aac5d3