Bounds inference querying does multiple passes if necessary.
This can only occur in one of two situations:

A) You're also inferring bounds on the output

B) The algorithm relies on the dimensions of the inputs
in a way that sets up a graph of constraints.

Unless you're in one of those situations, explicit constraints can't
trigger this, because in bounds inference mode those constraints
are applied to the inferred sizes, not the sizes passed in, so there's
no way for one pass of bounds inference to propagate information to
the next. The sizes passed in are ignored.


Former-commit-id: a13de8ef7c480f5184803f17b7a60f4e2cce3fbd