Fix typos.
Make memoizing a Func with different compute and store levels an error
per conversation with Andrew. (It is hard to make sense of such a
thing, though I had some ideas on cases where it might be useful. It
causes errors with the cache lookup being moved to be in the
conditional expression for a conditional allocation and when that
happens, the cache key storage is not in scope and compilation fails
with a symbol lookup error. I may debug this a bit more to make sure
it only happens when storeage and compute are scheduled to different
places.)