This change finishes plumbing through MIPS code generation support.
The HelloAndroid application is extended to build for multiple
architectures, including ARMv6, ARMv7a, 64-bit ARM, 32-bit MIPS,
32-bit x86 and 64-bit x86. (64-bit MIPS is not yet supported.)

GPU are added templates for all architectures. E.g. so GLSL can be
used from PNaCl and on mips

The usage of WITH_ preprocessor macros has been changed to consistently use #ifdef, not #if.

Added a README for building the HelloHalide app.


Former-commit-id: 2a2cc581b57a4fbe8c63ae0c89bf897bf779a144