Initial pass at removing varargs from the runtime.
This is necessary to avoid calling convention woes on nacl and windows.

Requires reworking AssertStmts to just take a string, so it touches a
lot of code.

Also required removing uses of halide_printf from the runtime, which
touches a lot of code there. It has been replaced with a std::cout-style
thing, which more closely matches how debugging prints work in the compiler.

This involved an implementation of float and double printing *shudder*.
Still, now we have platform-independent printing, and can remove the
windows-specific float printing hack too.

There are still a bunch of debug_prints in the runtime, with temporary
definitions of halide_printf and halide_error_varargs to make them work.
Fixing this is next.


Former-commit-id: 385ca80e134478d90caab7a9cba300878764d3cd