Changes to get shared_runtime and gpu_api_naming branches integrated and working.
There appears to be a bug with specialize tests and
InjectHostDevBuffercopies.

One remain task here is to figure out how to make the device_interface
API available in libHalide.a directly so that Buffer.cpp can get rid
of the soruce module support and just call halide_device_free
directly. At that point, Buffer and Image<T> can likely become free of
the main C++ Halide codebase so they can be used as RAII classes in
code that uses Halide AOT compiled filters.


Former-commit-id: c4a2ca9b31119cf58c75d8b3776b72489ce8516d