Add some better error checking for caser where the schedule specifies a given GPU API but it does not appear in the target. Before, this seg faulted.
Fix a typo in test_gpu_multi_device which was causing the wrong GPU
API to be selected, hence causing the above error.

Change the test used for "weak" in the heuristic to identify exported
Halide runtime routines from mayBeOverridden to isWeakForLinker. (This
is the umpteenth time this has changed. Pretty clear this is iffy...)

Remove now unecessary free_dev_buffer call in test_gpu_multi_device.