Fix a flaw in the IR equality cache hash
For some reason it has only showed up in this branch on 32-bit os x.

The hash was based purely on a difference (xor) of two addresses, which
means that you get the same hash when you flip the same bits in both addresses. I made it depend also on the sum of the addresses.

Former-commit-id: 88ec0085cc52e0fe0fa6bb264e57146ba3bec621