Improve performance when #threads > #tasks
While working on linear algebra, which occasionally uses more threads
than there are tasks, I noticed that a ton of time was spent in the
linux kernel from contention on the mutex. A pathological case is:

for i from 0 to 1000000:
parallel for j from 0 to 2:
    do something trivial

When a job comes in that has fewer tasks than threads, this commit puts
some of the threads into a deeper sleep, and only wakes them when a new
job comes in that can actually make use of them.

The windows thread pool is so far unchanged. OS X should hopefully not
suffer from this due to using gcd.


Former-commit-id: b1512748a5f77e5c8984bec597940a0b6fa0f1b4