Make buffer_t GPU routines such as halide_device_malloc/halide_device_free available directly from libHalide.a .
The above allows removing source module support from Buffer and
calling the halide_device routines directly.

For Target::JIT to be set on all shared runtime requests as the shared
runtime will not be made properly without that flag.

Couple tiny changes to OpenCL to ensure all platforms are considered
if a specific platform set is not given via environment variable and
to output the error number on the first cl call in case it is an
unknown error.

Remove source module support