Add halide_type_code_t
We need a runtime-suitable enum that corresponds to Type::TypeCode,
both for use in Metadata (Issue #664) and to add to buffer_t at some
point in the future. This defines an enum in HalideRuntime.h to pave
the way for these uses, and redefines the values of Type::TypeCode to
match them.

(It seems infeasible to replace Type::TypeCode entirely, as the runtime
version needs to be C-compatible, and preserving source compatibility
would require polluting the global namespace with (e.g.) “Int”, etc.)