[18496] - Implement a testing system for running unit tests.
This patch implements a system that uses a combination of a shell
script and lit/FileCheck to allow us to run small unit tests.
The shell script reads a 'testlist' file and compiles the tests
mentioned in that file using g++. Then lit and FileCheck are used
to run the executables so created - which are cross compilers
themselves, in one sense. The idea is that with every visit function
added, we'll create a corresponding test that will test the
generation of the command intended to be generated using the visitor
function.

Change-Id: I6596d26badc70dce6301339880c56584aab57dee
Reviewed-on: https://review-hexagon.quicinc.com/27238
Tested-by: Pranav Bhandarkar <pranavb@codeaurora.org>
Reviewed-by: Pranav Bhandarkar <pranavb@codeaurora.org>