Remove buffer_t.h
In favor of just including HalideRuntime.h, which also defines buffer_t,
and several other useful structs (e.g. tracing).

Including prototypes for runtime functions which don't exist in
libHalide is questionable (see the discussion in PR #675), but at
least it'll stop us from declaring prototypes with the same names
that conflict with the runtime versions.