Argument should distinguish Input and Output Buffers
Note that this introduces a potential change in GPU backends: formerly,
buffers were only prohibited from being read+write in the GLSL backend
(but not in other GPU backends); now it’s prohibited at the GPU_Host
level, so if other backends rely on read/write buffers, this would
break them. (I’m not sure how to test this.)

Also note that I replaced “Buffer::operator Argument” with
Buffer::as_input(), to emphasize that the Argument produced is always
an InputBuffer.