Make boundary conditions free
Involved adding a 'likely' intrinsic to tag the steady-state away from
the boundaries, and then making specialize_clamped_ramps (which needs to
be renamed) look for it and solve for loop variable mins and maxes that
simplify away code surrounding it.

Also used the likely intrinsic for sliding window, and for push the last
vector in a split inwards.

Interacts poorly with specialize_branched_ramps, so that is disabled for
now.