Fix bug with strided loads
There was a test for either generating slow safe strided loads for
external allocations, or fast slightly unsafe loads for internal
allocations. 'safe' here means whether it's ok to read slightly beyond
the end of an allocation.

Except the comparison was the wrong way around, so we were always doing
the slow thing for internal allocations, and the unsafe thing for
external ones. Doh. This speeds up local laplacian.


Former-commit-id: 7689dd216075929aeb4ac62328ade3494b955cf8