Add a "destructors" runtime module
Assists with registering and calling destructors. Makes it easy to call
all destructors on assertion failure without enumerating the ones in
scope, which saves a small amount of code size (several K after
stripping).

device_free (which was not previously called on assertion failure)
becomes a destructor on the buffer_t object. This part is pretty
awkward.


Former-commit-id: ee2c4c787890c82b1c5dd61bef2e4d66f82f61ff