Add scalar_from_constant_expr/scalar_to_constant_expr utilities
This was originally part of a PR for a fix for Issue #664, but it
looked fiddly enough that I thought it merited a separate PR on its own.

Basically, Parameter (and thus Argument) use a subset of Expr to hold
the default/min/max values for Generator params; having code to convert
between the Expr-wrapped forms (vs the actual-values we need to create
the metadata) encapsulated in one place makes the rest of the code more
reviewable.

(I’m not entirely happy with this code, as it feels more awkward than
it should be, but it will get the job done.)


Former-commit-id: 7e33aeb430b08a984f03f598cf76f46b3b92c2d2