A simpler solution to destructors
Allocates a stack void * per thing that needs a destructor and null
initializes it. Allocating adds a pointer to that stack slot. Freeing
the thing normally sets that pointer to NULL. Failing an assertion frees
all of the non-NULL ones in a single block of code.

This saves just over 2k in the final binary for local laplacian. Not a
huge win, but a win to be sure, and it provides a mechanism for
destructors in general in codegen.


Former-commit-id: 6f2a751d9a3c9267abc95de92d64552a7e173329