Change how error messages happen
Assert now evaluates and returns the second argument on failure. This
second argument must be an int.

All the existing places we were generating asserts now have second
arguments that call into error handlers in the runtime.

Cases which caught and reraised an error (e.g. errors in parallel for
loops triggered a second error in the calling function) now silently
pass the inner error outwards.

This moves the string construction code out of each filter and into the
runtime. Combined with the changes with destructors, this saves ~500
bytes or so per filter (and adds ~1500 bytes to the runtime).

The larger benefit though is likely to be meaningful error codes, and
the ability to catch and handle certain errors without parsing a string.

Still TODO: Actually make an enum for the error messages in
HalideRuntime.h, and change the destructor block to pass out the error
code instead of dropping it on the floor. Also need tests for overriding
the various error codes, and checking that the expected error code is
returned in each case.