Add optional metadata registration
When built with Target::RegisterMetadata (which is off by default), we
add a small amount of load-time code to build a list of information
about filter metadata and argv-style calls; this can be enumerated to
introspect/call filters.

When this Target feature is disabled, the only extra code size hit
added is the code in runtime/metadata.cpp (some trivial list
manipulation code); in particular, there are no hard references to
metadata, so it can still be dead-stripped.

When this feature is enabled, we add a trivial amount of code (a
ctor-ish function for each filter), but effectively prevent the
metadata structures from being dead-stripped, hence the opt-in nature
of this feature.

Please consider the TODO in HalideRuntime.h (i.e.: should we move
‘name’ and ‘argv_func’ into the metadata itself?)


Former-commit-id: 6ca020e5a36345f899ece0e67f1fb6b051473f15