Move metadata module to shared jit runtime
It's no longer inline-only stuff, so it shouldn't be included in every
jitted filter.

Required breaking a self-referential type due to our method for copying
a type from one llvm module to another not handling them.

This was prompted by a memory corruption bug I still don't understand -
setting the linkage type on halide_mutex_lock (used in metadata and also
in the shared runtime) in make_externs did an out of bounds read/write.
It should only be extern declared in the filter, and should actually
appear in the shared runtime. I'm not sure what the problem was. Moving
the metadata stuff to the shared runtime was correct anyway, but this
still worries me.


Former-commit-id: eed096aa4520f4fc23141e2b5611fabb84a7c7ee