Add a debugging GeneratorParam to all Generators
It’s handy to be able to leave debugging instrumentation in a Generator
without affecting real code; rather than using #ifdef or similar, let’s
add a standard GeneratorParam to enable this at the build-file level.

This defines it as a four-level enum; we could add more or less but
this seems likely to be a good compromise for most purposes.

(Unit test added to metadata_tester to minimize build-file churn.)