Generally working version of passing explict externs into JIT calls.
Need to test the Func path of JITExtern.

The variadic template code to destructure a function pointer at the
top of JITModule.h needs to be cleaned up. (E.g. can it be moved into
the JITExtern scope.)

Is JITModule.h the right place for JITExtern and is ScalarOrBufferT a
more general thing?

Need to write doc comments.


Former-commit-id: 6363a789fcdbb48c15f9ccc8ecd61eb3cc6fd1e8