Compile runtime at O3 to ensure elimination of dead globals
This is necessary to ensure that dead globals are eliminated - inline
linkage things become comdats, which breaks cross compilation to OS X.
The default set of optimization passes differs on Windows and Linux, so
I'm just turning on all optimizations on both platforms. CMakeFile
change to follow in next commit.