[19285] - Implement the Sobel filter in Halide.
This patch adds the sobel filter in Halide. It builds on the first
implmentation by Erich Plondke. To generate this, we had to implement
the absdiff instruction and the vmpyi - vector multiply with scalar.
This patch also adds a directory test/hexagon/apps which is now
the designated location for all Halide filters that we'll implement.

Change-Id: Id00b5fb3d90dba7293cef62b5b5433299854c454
Reviewed-on: https://review-hexagon.quicinc.com/28688
Reviewed-by: Ron Lieberman <ronl@codeaurora.org>
Tested-by: Ron Lieberman <ronl@codeaurora.org>