opengl.cpp: ensure that halide_get_device_handle() is never called with 0
It is an error condition to do so (but halide_get_device_handle() can’t
really do anything about it other than log a debug error and return
null); the caller should detect this and fail before making the call.
Also, some drive-by renaming to ensure that buf->dev isn’t confused
with the result of halide_get_device_handle().