Add test/platforms
This is a pull (and simplification) of code from abestephensg@ branch
https://github.com/halide/Halide/tree/simple-app-testing; it provides a
way to do simple Generator testing on iOS and OS X devices. Support for
Android devices will be added after this lands.

The intent here is to allow basic testing and performance metrics to be
gathered on mobile devices without a great deal of overhead.

(Note: the iOS code hasn’t been tested with most recent changes due to
a provisioning issue with the iOS device I have with me today, so it
may need modest attention.)


Former-commit-id: e186433b01be1ee925d85283a14bb24d15fcd653