Update realize with externs support for Pipeline. Move to an API where externs are set as state on a pipeline.
Introduced ExternSignature class so JITExtern can exist at Pipeline
level without its details being pushed down into JITModule.

Also, with Pipeline being strictly a pointer to a reference counted
object, which was not true with Func before, it can be treated as a
value by JITExtern and no longer has to be stored as a pointer.


Former-commit-id: 0ed09fdfd845a6fe0daca9d17b192da479e57849