Add Doxygen support to CMake build.
Unfortunately the ``Doxyfile`` needs to be configured (Doxyfile.in ->
Doxyfile) for out of source builds which means the Makefile also had to
be taught to do this.

An option ``WITH_DOCS`` has been added so that setting up the build
system to be able to build documentation can be disabled. Note the
``doc`` target is not run by default because it is a phony target and
if it was added to all it would run everytime.


Former-commit-id: 27b90caed5ddd6ea9f3e6a279a50bd3c9abfe031